+++ src/wrap.rs
@@ -8,20 +8,21 @@ use {
 #[derive(Debug, Default)]
 pub struct SubString {
     pub string_idx: usize,
-    pub byte_idx: usize,
+    pub byte_start: usize,
+    pub byte_end: usize, // not included
 }
 impl SubString {
     pub fn draw(&self, w: &mut W, report: &Report, line_idx: usize) -> Result<()> {
         let line = &report.lines[line_idx];
         let string = &line.content.strings[self.string_idx];
         if string.csi.is_empty() {
-            write!(w, "{}", &string.raw[self.byte_idx..])?;
+            write!(w, "{}", &string.raw[self.byte_start..self.byte_end])?;
         } else {
             write!(
                 w,
                 "{}{}{}",
                 &string.csi,
-                &string.raw[self.byte_idx..],
+                &string.raw[self.byte_start..self.byte_end],
                 CSI_RESET,
             )?;
         }
@@ -39,7 +40,7 @@ impl SubLine {
     pub fn is_continuation(&self) -> bool {
         self.sub_strings.get(0)
             .map_or(false, |sub_string| {
-                sub_string.string_idx != 0 || sub_string.byte_idx != 0
+                sub_string.string_idx != 0 || sub_string.byte_start != 0
             })
     }
     pub fn src_line<'r>(&self, report: &'r Report) -> &'r Line {
@@ -89,13 +90,24 @@ impl WrappedReport {
             let strings = &lines[line_idx].content.strings;
             for string_idx in 0..strings.len() {
                 let string = &strings[string_idx];
-                sub_lines.last_mut().unwrap().sub_strings.push(SubString { string_idx, byte_idx:0 });
+                sub_lines.last_mut().unwrap().sub_strings.push(SubString {
+                    string_idx,
+                    byte_start: 0,
+                    byte_end: string.raw.len(), // may be changed later on cut
+                });
                 for (byte_idx, c) in string.raw.char_indices() {
                     let char_cols = c.width().unwrap_or(0);
                     if sub_cols + char_cols > cols && sub_cols > 0 {
+                        sub_lines.last_mut().unwrap()
+                            .sub_strings.last_mut().unwrap()
+                            .byte_end = byte_idx;
                         sub_lines.push(SubLine {
                             line_idx,
-                            sub_strings: vec![ SubString { string_idx, byte_idx } ],
+                            sub_strings: vec![ SubString {
+                                string_idx,
+                                byte_start: byte_idx,
+                                byte_end: string.raw.len(), // may be changed later on cut
+                            } ],
                         });
                         sub_cols = char_cols;
                     } else {